feat(GUI): display a loading page when updating
We add an updating overlay that should only be shown during updates for
the Etcher Pro. This can be set using the `updatingState` field in the
Redux store, and setting it to `true`, and `false` when finished. We may
extend this to become an object containing more metadata in the future.

Closes: https://github.com/resin-io/etcher/issues/2312
Change-Type: patch